+++ CHANGES.md
@@ -17,6 +17,7 @@
   * Vars (created with `clojure.core/defn`) implement this protocol as an implementation detail
 * [BREAKING CHANGE] Now `dime.core/inject` is a fn and works with injectables only
 * [BREAKING CHANGE] Var-related API moved to `dime.var`
+* A fn `dime.core/deps-all` to extract dependency-keys from a dependency graph
 
 
 ## 0.1.0 / 2016-June-23